+++ src/Player/PlayerContinuousJump.cpp
@@ -12,7 +12,7 @@ void PlayerContinuousJump::update(bool shouldCountDown) {
 
     if (mCount > 0) {
         mTimer++;
-        if (mTimer >= mConst->getContinuousJumpTimer())
+        if (mTimer >= (u32)mConst->getContinuousJumpTimer())
             clear();
     }
 }
